requirements.txt
flask==2.3.3
openai==0.28.1
python-dotenv==1.0.0 